create base class for flashback-like abilities
